requirements.txt
mne
scikit-learn
numpy
matplotlib
pytorch-lightning
ml-collections==0.1.1
torchmetrics
pandas
seaborn